requirements.txt
websockets
ta
numpy
python-telegram-bot
blinker